pyproject.toml
[project]
name = "image-picker"
version = "0.1.0"
description = "CLI tool for scoring local photography images with LAION aesthetics and pyiqa."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "numpy>=1.26",
  "open-clip-torch>=2.24.0",
  "pandas>=2.2.0",
  "pillow>=10.3.0",
  "pyiqa>=0.1.13",
  "rawpy>=0.25.0",
  "scikit-learn>=1.5.0",
  "torch",
  "torchvision",
  "tqdm>=4.66.0",
  "typer>=0.12.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[project.scripts]
score-images = "image_picker.cli:main"
export-buckets = "image_picker.cli:export_main"
train-calibrator = "image_picker.cli:train_calibrator_main"
apply-calibrator = "image_picker.cli:apply_calibrator_main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/image_picker"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
